Add Unit test for ML inputs + Functional test for ML classes + python 2->3 conversion for validation.py + index problem for saving weights for datasurface
Former-commit-id: 337916f087591d66f3c59785c0e07599896a0100

+++ clinica/pipelines/machine_learning/input.py
@@ -386,11 +386,11 @@ class CAPSVertexBasedInput(CAPSInput):
 
         infinite_norm = np.max(np.abs(weights))
 
-        left_hemi_data = np.atleast_3d(np.divide(weights[:weights.size / 2], infinite_norm))
+        left_hemi_data = np.atleast_3d(np.divide(weights[:np.int(weights.size / 2)], infinite_norm))
         left_hemi_mgh = nib.MGHImage(left_hemi_data, affine=sample.affine, header=sample.header)
         nib.save(left_hemi_mgh, os.path.join(output_dir, 'weights_lh.mgh'))
 
-        right_hemi_data = np.atleast_3d(np.divide(weights[weights.size/2:], infinite_norm))
+        right_hemi_data = np.atleast_3d(np.divide(weights[np.int(weights.size/2):], infinite_norm))
         right_hemi_mgh = nib.MGHImage(right_hemi_data, affine=sample.affine, header=sample.header)
         nib.save(right_hemi_mgh, os.path.join(output_dir, 'weights_rh.mgh'))
         pass
